feat: unification of "Product List (Manual)" with the other CMS Product Lists (#673)
- renamed CMS component "Product List" to "Product List (Manual)"
