+++ src/app/shared/cms/cms.module.ts
@@ -9,7 +9,7 @@ import { CMSImageComponent } from './components/cms-image/cms-image.component';
 import { CMSLandingPageComponent } from './components/cms-landing-page/cms-landing-page.component';
 import { CMSProductListCategoryComponent } from './components/cms-product-list-category/cms-product-list-category.component';
 import { CMSProductListFilterComponent } from './components/cms-product-list-filter/cms-product-list-filter.component';
-import { CMSProductListComponent } from './components/cms-product-list/cms-product-list.component';
+import { CMSProductListManualComponent } from './components/cms-product-list-manual/cms-product-list-manual.component';
 import { CMSStandardPageComponent } from './components/cms-standard-page/cms-standard-page.component';
 import { CMSStaticPageComponent } from './components/cms-static-page/cms-static-page.component';
 import { CMSTextComponent } from './components/cms-text/cms-text.component';
@@ -71,7 +71,7 @@ import { SfeAdapterService } from './sfe-adapter/sfe-adapter.service';
       provide: CMS_COMPONENT,
       useValue: {
         definitionQualifiedName: 'app_sf_base_cm:component.common.productListManual.pagelet2-Component',
-        class: CMSProductListComponent,
+        class: CMSProductListManualComponent,
       },
       multi: true,
     },
