+++ src/app/shared/cms/components/cms-product-list-manual/cms-product-list-manual.component.html
@@ -1,0 +1,11 @@
+<div *ngIf="productSKUs?.length" class="product-list-container" [ngClass]="pagelet.stringParam('CSSClass')">
+  <h2 *ngIf="pagelet.stringParam('Title')">{{ pagelet.stringParam('Title') }}</h2>
+  <ish-products-list
+    [productSKUs]="productSKUs"
+    [listStyle]="pagelet.stringParam('ListStyle')"
+    [slideItems]="pagelet.numberParam('SlideItems')"
+    [listItemStyle]="pagelet.stringParam('ListItemStyle')"
+    [listItemCSSClass]="pagelet.stringParam('ListItemCSSClass')"
+  >
+  </ish-products-list>
+</div>
