--- src/app/shared/cms/components/cms-product-list/cms-product-list.component.html
@@ -1,14 +1,0 @@
-<div *ngIf="productSKUs.length" class="product-list-container" [ngClass]="pagelet.stringParam('CSSClass')">
-  <h2 *ngIf="pagelet.hasParam('Title')">{{ pagelet.stringParam('Title') }}</h2>
-  <div class="product-list row">
-    <div *ngFor="let sku of productSKUs" class="product-list-item" [ngClass]="pagelet.stringParam('ListItemCSSClass')">
-      <ish-product-item ishProductContext [sku]="sku" displayType="tile"></ish-product-item>
-    </div>
-  </div>
-</div>
-
-<!-- TODO: (CMS Component Configuration)
-  ProductView
-  ListStyle
-  SlideItems
--->
